services/booking/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
redis
aio_pika
httpx
grpcio
grpcio-tools
protobuf
